services/llm/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "AI Interview LLM Service (gRPC)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai==2.14.0",
    "redis==5.0.1",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-health-checking>=1.68.0",
    "protobuf>=5.28.0",
    "python-dotenv==1.0.0",
    "edge-tts>=6.1.12",
]

[tool.uv]
package = false
